Touch a different file with each test
To ensure that each test iteration is measuring the correct thing, I
rework the tests so that instead of appending random content to
AkkaMain.scala, I add another file WatchFile.scala that contains the
specifies which file to write the output to. For each iteration, we
rewrite that file and add a PathWatcher callback that only triggers for
the specific file we're interested in and only when the content matches
the expected content.