Add support for bloop
This was a bit tricky because bloop must run as a a client/server
pair so I had to add a mechanism for potentially managing multiple
forked process for a given build tool. The setup of bloop is also more
involved since the build is designed to be exported. It doesn't do its
own dependency resolution so eliminating the sbt bloopInstall step is
difficult.